perf(wishlist): ditch PImpl and make methods constexpr
